ai_agent_microservice/requirements.txt
aiofile==3.9.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1753980460404/work
anaconda-auth @ file:///croot/anaconda-cloud-auth-split_1747863777058/work
anaconda-cli-base @ file:///croot/anaconda-cli-base_1741369443553/work
annotated-doc==0.0.4
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
anyio==4.13.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==26.1.0
Authlib==1.7.0
beartype==0.22.9
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==7.0.5
caio==0.9.25
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1754477934456/work
colorama @ file:///work/perseverance-python-buildout/croot/colorama_1728388434940/work
conda @ file:///croot/conda_1754469510968/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1755123313635/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
cryptography @ file:///croot/cryptography_1753084572505/work
cyclopts==4.10.2
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
dnspython==2.8.0
docstring_parser==0.18.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.136.0
fastmcp==3.2.4
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
git-filter-repo==2.47.0
graphifyy==0.4.14
griffelib==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
importlib_metadata==8.7.1
jaraco.classes @ file:///croot/jaraco.classes_1755516327832/work
jaraco.context @ file:///io/perseverance-python-buildout/croot/jaraco.context_1731706750665/work
jaraco.functools @ file:///croot/jaraco.functools_1740408111753/work
jeepney @ file:///tmp/build/80754af9/jeepney_1627537048313/work
jiter==0.14.0
joblib==1.5.3
joserfc==1.6.4
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer @ file:///croot/jsonpointer_1753788433671/work
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring @ file:///croot/keyring_1745484691218/work
langchain-core==1.3.0
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.32
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
mcp==1.27.0
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1753464582829/work
more-itertools @ file:///work/perseverance-python-buildout/croot/more-itertools_1728388668839/work
-e git+ssh://git@github.com/aryan68125/Music-Player.git@b70fdda040231ec242f2a9e9345247b982161dd4#egg=music_player
mutagen==1.47.0
networkx==3.6.1
nltk==3.9.3
openai==2.32.0
openapi-pydantic==0.5.1
opentelemetry-api==1.41.0
orjson==3.11.8
ormsgpack==1.12.2
packaging @ file:///croot/packaging_1753775359799/work
pathable==0.5.0
pkce @ file:///work/perseverance-python-buildout/croot/pkce_1728388093137/work
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
py-key-value-aio==0.4.4
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic-settings @ file:///io/buildout/croot/pydantic-settings_1735843225507/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
Pygments @ file:///croot/pygments_1744664109463/work
PyJWT @ file:///io/buildout/croot/pyjwt_1739558969999/work
pyperclip==1.11.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.11.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
python-multipart==0.0.26
PyYAML==6.0.3
readchar @ file:///work/perseverance-python-buildout/croot/readchar_1728397613245/work
referencing==0.37.0
regex==2026.2.28
requests @ file:///croot/requests_1750426329888/work
requests-toolbelt==1.0.0
rich @ file:///croot/rich_1732638981168/work
rich-rst==1.3.2
rpds-py==0.30.0
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
SecretStorage @ file:///work/perseverance-python-buildout/croot/secretstorage_1728387678881/work
semver @ file:///croot/semver_1736540782504/work
setuptools==78.1.1
shellingham @ file:///work/perseverance-python-buildout/croot/shellingham_1728395303315/work
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tomli @ file:///croot/tomli_1753774587605/work
tqdm @ file:///croot/tqdm_1738943501192/work
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-c-sharp==0.23.5
tree-sitter-cpp==0.23.4
tree-sitter-elixir==0.3.5
tree-sitter-go==0.25.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-julia==0.23.1
tree-sitter-kotlin==1.1.0
tree-sitter-lua==0.5.0
tree-sitter-objc==3.0.2
tree-sitter-php==0.24.1
tree-sitter-powershell==0.26.3
tree-sitter-python==0.25.0
tree-sitter-ruby==0.23.1
tree-sitter-rust==0.24.2
tree-sitter-scala==0.24.0
tree-sitter-swift==0.0.1
tree-sitter-typescript==0.23.2
tree-sitter-verilog==1.0.3
tree-sitter-zig==1.1.2
truststore @ file:///croot/truststore_1753907171582/work
typer @ file:///work/perseverance-python-buildout/croot/typer_1728397851651/work
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.3.1
urllib3 @ file:///croot/urllib3_1750775463400/work
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
wheel==0.45.1
xxhash==3.6.0
zipp==3.23.1
zstandard @ file:///croot/zstandard_1731356346222/work

# ── auto_classifier agent dependencies ──────────────────────────────────────
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pgvector>=0.3
redis>=5.0
python-jose[cryptography]>=3.3
structlog>=24.1
lxml>=5.2
httpx>=0.27
